backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
mysql-connector-python
python-dotenv
Flask-JWT-Extended
Werkzeug
Flask-Marshmallow
marshmallow-sqlalchemy
Flask-CORS